Add files via upload
Source code for KraftWish so you can go through the code and see if there is any viruses or malware if you are skeptical. Or maybe if you wanna check if when you enter your password, I'll just steal it. So check if you'd like, or run it through virustotal, I assure you this app is meant for an assignment for school, not an app used for malicous intent. But by all means, go ahead and check the source code.